Move DailyReminderDebugPreferences to main
