+++ mobile/src/debug/java/com/alexstyl/specialdates/debug/DebugFragment.java
@@ -14,6 +14,7 @@ import android.widget.Toast;
 
 import com.alexstyl.specialdates.R;
 import com.alexstyl.specialdates.contact.actions.IntentAction;
+import com.alexstyl.specialdates.dailyreminder.DailyReminderDebugPreferences;
 import com.alexstyl.specialdates.date.DayDate;
 import com.alexstyl.specialdates.service.DailyReminderIntentService;
 import com.alexstyl.specialdates.ui.base.MementoPreferenceFragment;
